Do not raise an error when a DS9 command fails.
As the idea is to run the calls within a context generator then
we don't want the code to raise an error if there's a syntax error
when calling ds9.set or ds9.get, as that would then cause the
connection to be closed, annoying the user.

So, instead we just print a string message if there is a problem.
These messages are to stdout and not stderr as it doesn't seem to
be worth the effort to make them go to that channel.

Note that the calls can still raise an error, but that is from
a SAMP-related problem, which we probably want to exit the loop.