Add the retrieve_array method
Since SAMP doesn't really send data (other than URLs) in band,
we write the array out and read it back in via NumPy memmap.

I am not 100% happyt with this, as the NumPy memmap documentation
discusses lifetime issues, but let's see how this works.

An alternative (that could be done with the same API) is to
write out a FITS file and then load that in. However, that is
for the future *IF* this approach turns out to be problematic.