Add the mask argument to send_array
For "future-proofing", make the send_array and retrieve_array
calls require all optional arguments as keyword-only. Perhaps
this should be done for get and set too, but for now resist this
impulse.

As part of this, since FITS doesn't have a "boolean" type, convert
boolean images to int8 as it hopefully wastes the least memory.