Allow 3D arrays to be sent and retrieved.
As well as supporting 3D arrays, the code has seen a minor update
to better distinguish between dealing with problems with data the
user has sent - which generally means raising an error - and with
data from DS9 - which generally means printing a message and
skipping the rest of the action.

As part of this, move the "add colour" routine from scripts into
the main code and make use of it. This is somewhat experimental
as I may want to switch to a proper logging library, in which case
the colour shenanighans would be turned off for the library code.