Add the debug option to the connection.
Let's see whether this is useful and whether we want to do
something like display the return value ni a more-structured
form.

I thought about making debug be a field in the constructor,
but as this is not meant for general-purpose use we require the
caller to manually set the field after creating the object.