Retrieve data sent via the url response
get calls can return data via the `url` field of the result,
rather than `value`. If this happens, grab the contents and
return it to the user.

There is a minimal check that the url is as expected. An
alternative would be a proper URL parser, which might be useful
if DS9 were to take advantage of it, but it doesn't seem worth
it right now.