Add the send/retrieve_fits methods
This work shows that the existing break down of code needs to
be re-worked (e.g. perhaps have _get/_set methods which just
return the SAMP result after minor processing?).

It's also not clear what the best interface here is, since it's
not obvious that something like the Cube enumeration is useful,
but more experience is needed.