Add the send_cat and retrieve_cat methods
The choice of FITS as the method to send data to DS9 and
TSV for the return values are internal to the code, so may
be changed at any time. However, AstroPy 7.0.0 did not
appreciate the RDB dat athat was created by DS9 8.7 beta 1,
so there are some limitations.

It is not best clear what the most-useful interface is here.
For example, we could send some form of a NumPy recarray.
However, I think it best if we try to leave the format conversions
to user code and try to limit the complexity here.