Refactor code added in 0.0.7
When writing retrieve_fits I noted that it repeated code we already
had. So we create a _get method which calls ds9.get and does
basic error checks on the response, but does not try to interpret
the "samp.result" field. The get() and retrieve_fits() can then
apply their own interpretation.